The expression return mysqli_affected_rows($this->conn) also could return the type string which is incompatible with the documented return type integer.

+++ htdocs/class/database/mysqldatabase.php
@@ -176,7 +176,7 @@ abstract class XoopsMySQLDatabase extends XoopsDatabase
      */
     public function getAffectedRows()
     {
-        return mysqli_affected_rows($this->conn);
+        return (int)mysqli_affected_rows($this->conn);
     }
 
     /**
